docs: update README with new features and version-specific attributes
- Document V3-only attributes (overflow, linebreakmultchar)
- Document content elements support (cbytes, cs, csymbol with presentation elements)
- Document supported Content MathML elements
- Update Unsupported Features table with new HTML attributes and foreign content
- Add Known Pending Issues section for lutaml-model and parser limitations

+++ lib/mml/base/mscarries.rb
@@ -7,7 +7,7 @@ module Mml
       # Use fully qualified names (e.g., Mml::Namespace).
       def self.included(klass)
         klass.class_eval do
-          attribute :scriptsizemultiplier, :integer
+          attribute :scriptsizemultiplier, :float
           attribute :mathbackground, :string
           attribute :mathcolor, :string
           attribute :position, :integer
